RePlayerSFX: Add second sound effects channel.
It‘s probably a good convention to:
* Use channel 1 for sound effects originating from player.
* Use channel 2 for sound effects not originating from player.

This should reduce sound effects cutting each other off.
An additional channel could be added if necessary.

Two identical copies of the sound effects are included, the only difference
being which OPL4 FM channel they play on. The appropriate one is picked based
on which SFX stream channel the sound is played on.