Usas2: Fix subslot issue when selecting RAM in page 2.
The engine uses a trick to avoid secondary slot selection. This method requires
the subslot register to be initialised once in initMem. However, it did not do
this for the RAM in page 2.

This caused intermittent issues depending on machine configuration, boot order,
etc. For example the area sign would sometimes show up as garbage. It tries to
select RAM in page 2, but if RAM was not selected it can not unpack the image.
The world map and F1 / F2 menus probably also had similar issues.

This fix adds a call to ENASLT (24H) for page2ram so that the secondary slot
register gets pre-configured.

Also add comments to better describe the slot selection method.